refactor(test): rename a utl to .utl so it's easier to distinguish
